+++ apps/web/src/app/api/ingest-posts/route.ts
@@ -1,7 +1,7 @@
 import type { NextRequest } from "next/server";
 import db from "../../db";
 import { convertPostToProductPost, getAllPost, getAllPostsVotesMoarBetter } from "../../lib/data";
-import { batchAnalyzePosts } from "../../services/ai-analyzer";
+import { analyzePosts } from "../../lib/ai-analyzer";
 import { PRODUCT_HUNT_NAME } from "../../utils/string";
 import type { Post, PostResponse, ProductPost } from "../../types";
 
@@ -38,7 +38,7 @@ export async function GET(request: NextRequest) {
     ...convertPostToProductPost(post),
     id: post.id,
   }));
-  const aiAnalysisResults = await batchAnalyzePosts(postsToAnalyze);
+  const aiAnalysisResults = await analyzePosts(postsToAnalyze);
 
   // Create a map of post IDs to their AI analysis results
   const postAiResults = new Map(
